refactor kverrorhandler
